Add support for ticket systems which only use digits for IDs
Note that a part of the commit message might undesirably be replaced by the ticket ID if your commit message starts/ends with a number and matches the prefix settings. A test that illustrates this has been included.

Also, when switching between issue-tracking systems, an additional prefix might be added. This, too, is illustrated with a test.